requirements.txt
Flask==2.3.2
flask-cors==4.0.0
Jinja2==3.1.3
Gunicorn==21.2.0
unidecode==1.3.6
nltk==3.8.1
scikit-learn==1.3.0
spacy==3.6.0
torch==2.1.0
werkzeug==3.0.1
requests==2.31.0
pycodestyle==2.11.1
python-dotenv==1.0.0
flasgger==0.9.7.1
waitress==2.1.2
locust==2.21.0
